+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file. This projec
 - Implement better icons for the app
 - Cleanup of Disk analysis code
 - Improved UI of disk analysis page
+- Added progress bar for File explorer
 
 ## 0.9.9
 - Fix issue where some popover components are transparent